Added API call for auditor to add audit details
Previously, we tried to send a PATCH request to /api/stock/id
as auditor to audit stocks and it failed since auditor does not
have the required permissions to modify stock. Changed it to use
a newly created API call which gives only permisssion to add an
AuditDetail object to a Stock.